chore: upgrade all dependencies
Note: the force-graph update comes with the ability to extend the force
graph types, by using generics. Using this feature, I just fixed the
whole typing issue we had, with callbacks, that receive graph data from
the force graph. This data was previously just typed with the
force-graph types, but now we are able to extend it.