Feat: smooth edit (#6)
Create a smooth in-graph editing experience.

- Ctrl+Click on background to add node, center on new node
- Drag nodes close to each other to link them together
- Big fat (+)-button lower-right corner for node and link creation
- Disable Zoom feature for now, it is not a good user experience combined with node & link creation

Small things:
- Add shortcuts for almost all editing steps, these are displayed onHover of the graphical edit menus
- Links get highlighted when hovering over them
- Add descriptive marks for the link-weight-slider

Technical:
Update force-graph dependency versions. It comes with the ability to extend the
force graph types, by using generics. Using this feature, it was possible to
fix the whole typing issue we had, with callbacks, that receive graph data from
the force graph. This data was previously just typed with the force-graph
types, but now we are able to extend it.